Implemented isometric view
Decided that isometric view is much better and more stylized. (thank god I tried it and got some opinions). Implemented this by making the camera orthographic and adjusted the camera positions. Now the camera is rotated to give a nice and clean isometric view. This should work very well with the assets we plan to use.

Next, we'll want to start building a simple UI. For now, use the old UI system for a lot of these functions. I want to experiment with having the UI within the level and ultimately want to go that way; I just need to figure out how to do that effectively for controllers.